+++ apps/admin/providers/dataProvider.ts
@@ -1,8 +1,4 @@
 import { DataProvider } from "react-admin";
 import { dataProvider as prismaDataProvider } from "ra-data-simple-prisma";
 
-export const dataProvider: DataProvider = prismaDataProvider("/api", {
-  resourceToModelMap: {
-    admin: "adminUser",
-  },
-});
+export const dataProvider: DataProvider = prismaDataProvider("/api");